# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital Shagun & Social Gifting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or get user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created or retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /shagun:
    post:
      operationId: sendShagun
      tags: [shagun]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendShagunRequest"
      responses:
        "201":
          description: Shagun sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /shagun/{eventId}:
    get:
      operationId: getShagunByEvent
      tags: [shagun]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of shagun for event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /shagun/reveal/{transactionId}:
    get:
      operationId: revealShagun
      tags: [shagun]
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reveal status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealStatus"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
      required: [id, name, phone, avatarColor, createdAt]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required: [name, phone]
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        receiverId:
          type: string
        amount:
          type: number
        message:
          type: string
        isRevealed:
          type: boolean
        revealAt:
          type: string
        createdAt:
          type: string
      required: [id, eventId, senderId, senderName, receiverId, amount, createdAt]
    SendShagunRequest:
      type: object
      properties:
        eventId:
          type: string
        receiverId:
          type: string
        amount:
          type: number
        message:
          type: string
      required: [eventId, receiverId, amount]
    RevealStatus:
      type: object
      properties:
        id:
          type: string
        isRevealed:
          type: boolean
        amount:
          type: number
        senderName:
          type: string
        message:
          type: string
        secondsRemaining:
          type: number
      required: [id, isRevealed, amount, senderName, secondsRemaining]
...